Change create_library behavior when library exists
Originally, create_library would raise an error if the library
already existed, but the requested headers differed from those
in use. The method behaved as you might expect in the other two
cases.

Now, if the headers change, then the library and headers tables
are dropped and recreated using the new headers. No error is
explicitly raised.

One problem at the moment, is that this new approach results in
the index in a clean state. It would be nice to be able to reindex
all of the directories that had previously been indexed.